Linux

+++ core-esp32/src/core/EventBus.h
@@ -18,6 +18,8 @@
 #include <esp_check.h>
 #include <esp_event.h>
 #include <list>
+#include <string>
+#include <string_view>
 #include "Logger.h"
 
 typedef uint16_t MsgId;
@@ -279,7 +281,7 @@ ESP_EVENT_DECLARE_BASE(CORE_EVENT);
 struct BusOptions {
     bool useSystemQueue{false};
     int32_t queueSize{32};
-    size_t stackSize{4096};
+    uint32_t stackSize{4096};
     size_t priority{tskIDLE_PRIORITY};
     std::string name;
 };
@@ -361,7 +363,7 @@ public:
             opt(options);
         }
 
-        esp_logi(bus, "create queue, size: %ld, item-size: %d", options.queueSize, itemSize);
+        esp_logi(bus, "create queue, size: %" PRIi32 ", item-size: %zu", options.queueSize, itemSize);
         _queue = xQueueCreate(options.queueSize, sizeof(Item));
 
         auto res = xTaskCreate(
